move Parse_Errors, Parse_Result, Error_Builder and Response_Builder under Core\Utils\Api
