+++ core/utils/api/parse-errors.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Elementor\Modules\AtomicWidgets\Parsers;
+namespace Elementor\Core\Utils\Api;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
